Pull out LoremIpsumGenerator to foundation module
