+++ library/foundation/src/main/java/com/epmedu/animeal/foundation/util/LoremIpsumGenerator.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.tabs.more.faq.util
+package com.epmedu.animeal.foundation.util
 
 import androidx.compose.ui.tooling.preview.datasource.LoremIpsum
 import kotlin.random.Random
@@ -7,7 +7,7 @@ import kotlin.random.Random
  * Generates Lorem Ipsum text with random length in specified [wordsCountRange]
  * @param wordsCountRange Words count range for randomizer. By default 5..100
  */
-internal fun generateLoremIpsum(wordsCountRange: IntRange = 5..100) =
+fun generateLoremIpsum(wordsCountRange: IntRange = 5..100) =
     LoremIpsum(
         words = Random.nextInt(
             from = wordsCountRange.first,
